auslagerung funktionalität in composable
