+++ frontend/src/components/messstelle/charts/BelastungsplanMessquerschnittCard.vue
@@ -14,6 +14,7 @@ import _ from "lodash";
 import LadeBelastungsplanMessqueschnittDataDTO from "@/types/messstelle/LadeBelastungsplanMessqueschnittDataDTO";
 import { useStore } from "@/api/util/useStore";
 import { useVuetify } from "@/util/useVuetify";
+import { belastungsplanAnzeigeUtils } from "@/components/messstelle/optionsmenue/composable/belastungsplanAnzeigeUtils";
 
 interface Props {
     belastungsplanData: ListBelastungsplanMessquerschnitteDTO;
@@ -41,7 +42,8 @@ const farben = new Map<string, string>([
 
 const startX = ref(0);
 const startY = ref(0);
-
+const { isSv_pInBelastungsPlan, isGv_pInBelastungsPlan } =
+    belastungsplanAnzeigeUtils();
 const props = defineProps<Props>();
 
 const svgHeight = computed(() => {
@@ -61,7 +63,6 @@ onMounted(() => {
 });
 
 function drawingConfig() {
-    console.log("config");
     canvas.value
         .addTo("#drawingMessquerschnittBelastungsplan")
         .size(svgHeight.value, svgHeight.value)
@@ -72,7 +73,6 @@ function drawingConfig() {
         .groupBy("direction")
         .map((value, key) => ({ direction: key, data: value }))
         .value();
-    console.log(groupedByDirection);
     startX.value = 150;
     startY.value = 250;
     draw();
@@ -84,54 +84,6 @@ const chosenOptionsCopy = computed(() => {
 
 const chosenOptionsCopyFahrzeuge = computed(() => {
     return chosenOptionsCopy.value.fahrzeuge;
-});
-
-const isSv_pInBelastungsPlan = computed(() => {
-    let actualNumberOfSelectedKfzSvAndGv = 0;
-    if (chosenOptionsCopyFahrzeuge.value.kraftfahrzeugverkehr) {
-        actualNumberOfSelectedKfzSvAndGv++;
-    }
-    if (chosenOptionsCopyFahrzeuge.value.schwerverkehr) {
-        actualNumberOfSelectedKfzSvAndGv++;
-    }
-    if (chosenOptionsCopyFahrzeuge.value.gueterverkehr) {
-        actualNumberOfSelectedKfzSvAndGv++;
-    }
-    return (
-        chosenOptionsCopyFahrzeuge.value.schwerverkehrsanteilProzent &&
-        (chosenOptionsCopyFahrzeuge.value.kraftfahrzeugverkehr ||
-            chosenOptionsCopyFahrzeuge.value.schwerverkehr ||
-            chosenOptionsCopyFahrzeuge.value.gueterverkehr) &&
-        actualNumberOfSelectedKfzSvAndGv < 3
-    );
-});
-
-/**
- * Hilfsmethode, um zu schauen, ob der Wert GV% im Belastungsplan angezeigt wird.
- * Dies ist nur der Fall, wenn KFZ, SV oder GV aktiviert sind und inklusive GV_P nicht
- * mehr wie 3 Verkehrsarten (ohne RAD und FUSS) ausgewählt sind
- */
-const isGv_pInBelastungsPlan = computed(() => {
-    let actualNumberOfSelectedKfzSvGvAndSV_P = 0;
-    if (chosenOptionsCopyFahrzeuge.value.kraftfahrzeugverkehr) {
-        actualNumberOfSelectedKfzSvGvAndSV_P++;
-    }
-    if (chosenOptionsCopyFahrzeuge.value.schwerverkehr) {
-        actualNumberOfSelectedKfzSvGvAndSV_P++;
-    }
-    if (chosenOptionsCopyFahrzeuge.value.gueterverkehr) {
-        actualNumberOfSelectedKfzSvGvAndSV_P++;
-    }
-    if (chosenOptionsCopyFahrzeuge.value.schwerverkehrsanteilProzent) {
-        actualNumberOfSelectedKfzSvGvAndSV_P++;
-    }
-    return (
-        chosenOptionsCopyFahrzeuge.value.gueterverkehrsanteilProzent &&
-        (chosenOptionsCopyFahrzeuge.value.kraftfahrzeugverkehr ||
-            chosenOptionsCopyFahrzeuge.value.schwerverkehr ||
-            chosenOptionsCopyFahrzeuge.value.gueterverkehr) &&
-        actualNumberOfSelectedKfzSvGvAndSV_P < 3
-    );
 });
 
 function drawTotal(querschnittGroup: any) {
